+++ src/runner/bin.ts
@@ -1,8 +1,9 @@
 import index from './'
 
+// Run hook
 index(process.argv)
-  .then(status => process.exit(status))
-  .catch(err => {
+  .then((status: number): void => process.exit(status))
+  .catch((err: Error): void => {
     console.log('Husky > unexpected error', err)
     process.exit(1)
   })